Moved `usePolling` from [packages/web/src/viewer/hooks/use-polling.ts](/Users/hal/Documents/VSCodeProjects/n-dx-internal/packages/web/src/viewer/hooks/use-polling.ts) to [packages/web/src/viewer/views/use-polling.ts](/Users/hal/Documents/VSCodeProjects/n-dx-internal/packages/web/src/viewer/views/use-polling.ts), updated the hook barrel and direct imports/tests, and removed the stale zone pin from [.n-dx.json](/Users/hal/Documents/VSCodeProjects/n-dx-internal/.n-dx.json).
