+++ .n-dx.json
@@ -10,7 +10,6 @@
       "mergeThreshold": 5,
       "pins": {
         "packages/web/src/viewer/views/hench-runs.ts": "web-dashboard",
-        "packages/web/src/viewer/hooks/use-polling.ts": "web-viewer",
         "packages/web/src/viewer/hooks/use-gateway.ts": "web-viewer",
         "packages/web/src/viewer/hooks/use-project-status.ts": "web-viewer",
         "packages/web/tests/unit/viewer/use-project-status.test.ts": "web-viewer",
