+++ .rex/prd.json
@@ -3663,20 +3663,24 @@
         {
           "id": "573db51d-2b91-4755-8910-48f0b123c2a8",
           "title": "Fix move-file in web-viewer (1 finding)",
-          "status": "pending",
+          "status": "completed",
           "level": "feature",
           "description": "- File \"packages/web/src/viewer/hooks/use-polling.ts\" is pinned to zone \"Web Viewer\" but lives in packages/web/src/viewer/hooks/ — consider moving to packages/web/src/viewer/views/ to align physical location with architectural zone",
           "priority": "high",
           "source": "sourcevision",
+          "startedAt": "2026-04-20T15:38:47.231Z",
+          "completedAt": "2026-04-20T15:38:47.231Z",
           "children": [
             {
               "id": "26cb2890-9483-4005-a6de-ef83ca613eed",
               "title": "Fix move-file in web-viewer: File \"packages/web/src/viewer/hooks/use-polling.ts\" is pinned to zone \"Web Viewe",
-              "status": "pending",
+              "status": "completed",
               "level": "task",
               "description": "- File \"packages/web/src/viewer/hooks/use-polling.ts\" is pinned to zone \"Web Viewer\" but lives in packages/web/src/viewer/hooks/ — consider moving to packages/web/src/viewer/views/ to align physical location with architectural zone",
               "priority": "high",
               "source": "sourcevision",
+              "startedAt": "2026-04-20T15:35:32.503Z",
+              "completedAt": "2026-04-20T15:38:47.177Z",
               "recommendationMeta": {
                 "findingHashes": [
                   "982af57cac3f"
