+++ packages/web/src/viewer/hooks/use-project-status.ts
@@ -16,7 +16,7 @@
  */
 
 import { useState, useEffect, useCallback, useRef } from "preact/hooks";
-import { usePolling } from "./use-polling.js";
+import { usePolling } from "../views/use-polling.js";
 import {
   createWSPipeline,
   isFeatureDisabled,
